get test groups and cases using the private `Invoker` class in `patrolTest()`
This place is not ideal, because when code inside of `patrolTest()` starts
executing, the test has, conceptually, already started.
We want to know about all groups and cases earlier.

It's also a bummer that Declarer doesn't have any public getter that'd provide
us with a view into the tests that are declared.