Find the better way to report Dart tests to the native side
We create a single, special test case (internal to Patrol). That test executes
as the first one, and inside of it, we explore the structure of Dart tests
and send them to the native side. The native side will be in charge of telling
the Dart tests to run.

Possible problems:
- when randomization of test case execution order is enabled, this will break