find a hack and make it work
Instead of: trying to "run all Dart tests at once", and immediately hold the
execution of all of them except for the requested one

We do: insert a check at the beginning of test execution that checks if this test
was requested by the native side. If not, the test early-returns and the next
one starts. Eventually, we'll find a test that was requsted, and run it.